backend/requirements.txt
# Backend Dependencies for GitLab GenAI Chatbot
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and RAG Dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.23
langchain-community>=0.0.12
langchain-groq>=0.1.0
openai>=1.6.0
anthropic>=0.8.1
groq>=0.4.0

# Vector Database
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Utilities
pandas>=2.2.0
numpy>=1.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS is built into FastAPI
